Skip TRS call to update to withdrawn if TRN blank
If a trainee is withdrawn immediately after registration TRS fails on
the call to update professional status because it requires a TRN to
process the update. We can avoid this by checking for the presence of a
TRN and skipping the call if a TRN is not present. There is already a
call to update professional status after we receive a TRN and this
should be sufficient to keep TRS in sync.